Module 7 Complete
In this module I revisited looping and swtich statements using the new syntax of 'var (variable) when (variable meets condition)'

I also revisited the try catch and error handling for the grade program. Then came across some new syntax with the 'finally' block, which will always execute regardless of exceptions caught. An example of the implementation of this is the closing of sockets or the protection of data, to prevent resources becoming compromised misused.